backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
SQLAlchemy>=2.0
pymysql
python-dotenv
python-dateutil
openai
faiss-cpu
numpy
python-multipart
cryptography
